Verify garden file doesn't exist

+++ src/app/beer_garden/db/mongo/util.py
@@ -158,21 +158,20 @@ def ensure_v3_24_model_migration():
         for legacy_garden in garden_collection.find():
             if legacy_garden["connection_type"] != "LOCAL":
 
-                garden_file_data = {"receiving": False, "publishing": False}
+                if not os.path.exists(config.get(f"{config.get('children.directory')}/{legacy_garden['name']}.yaml")):
+                    garden_file_data = {"receiving": False, "publishing": False}
 
-                if legacy_garden["connection_type"] == "HTTP":
-                    garden_file_data["http"] = legacy_garden["connection_params"]["http"]
-                if legacy_garden["connection_type"] == "STOMP":
-                    garden_file_data["stomp"] = legacy_garden["connection_params"]["stomp"]
+                    if legacy_garden["connection_type"] == "HTTP":
+                        garden_file_data["http"] = legacy_garden["connection_params"]["http"]
+                    if legacy_garden["connection_type"] == "STOMP":
+                        garden_file_data["stomp"] = legacy_garden["connection_params"]["stomp"]
 
-                logger.warning(f"Mapping Child Config: {config.get('children.directory')}/{legacy_garden['name']}.yaml")
-                logger.warning(f"Legacy={legacy_garden}")
-                logger.warning(f"Mapped={garden_file_data}")
-                with open(
-                    f"{config.get('children.directory')}/{legacy_garden['name']}.yaml",
-                    "w+",
-                ) as ff:
-                    yaml.dump(garden_file_data, ff, allow_unicode=True)
+                    logger.warning(f"Mapping Child Config: {config.get('children.directory')}/{legacy_garden['name']}.yaml")
+                    with open(
+                        f"{config.get('children.directory')}/{legacy_garden['name']}.yaml",
+                        "w+",
+                    ) as ff:
+                        yaml.dump(garden_file_data, ff, allow_unicode=True)
 
         db.drop_collection("garden")
